+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 ### Other changes
 
 * Fixed an issue where the user interface did not necessarily display distances in the same measurement system as the requested route by default.
+* Fixed the fallback to the `RouteOptions.locale` for on-device speech synthesizing if the Mapbox Voice API does not support the requested language.
 
 ## v3.8.0